ai/requirements.txt
# DeepAgents 依赖
deepagents>=0.1.0

# LangChain 核心依赖（deepagents 依赖）
langchain>=0.3.0
langgraph>=0.2.0

# Anthropic Claude API
langchain-anthropic>=0.3.0

# OpenAI 兼容 API (用于 DeepSeek)
langchain-openai>=0.3.0

# 环境变量管理
python-dotenv>=1.0.0

# SOCKS 代理支持（如果使用代理）
httpx[socks]>=0.28.0

# HTTP 服务（FastAPI）
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=2.0.0

# HTTP 客户端（用于测试）
requests>=2.31.0

# 可选：用于搜索功能
# tavily-python>=0.5.0
